+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Change to use underscores when the default destination folder name contains invalid characters.
+
+### Fixed
+- Fixed the issue where conversion fails when the destination folder name contains invalid characters.
+- Fixed the issue where conversion fails when the destination folder name ends with a space.
+
 ## [1.13.1] - 2023-09-12
 
 ### Fixed